requirements.txt
Django==4.0.4
djangorestframework==3.13.1
django-filter==21.1  # Recommended by Django REST framework
djoser==2.1.0  # REST implementation of Django authentication system
Markdown==3.3.6  # Recommended by Django REST framework
